+++ src/users/users.controller.ts
@@ -7,38 +7,43 @@ import {
   Patch,
   Delete,
   UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto, UpdateUserDto } from './dto';
 import { AuthGuard } from '@nestjs/passport';
+import { GetUser } from 'src/auth/get-user.decorator';
 
 @Controller('users')
 @UseGuards(AuthGuard('jwt'))
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
-  @Post()
-  create(@Body() userData: CreateUserDto) {
-    return this.usersService.createUser(userData);
-  }
-
   @Get(':id')
-  getUser(@Param('id') id: string) {
+  getUser(@Param('id') id: string, @GetUser('userId') userId: string) {
+    if (id !== userId) {
+      throw new ForbiddenException('Access denied');
+    }
     return this.usersService.getUserById(id);
   }
 
-  @Get()
-  getAllUsers() {
-    return this.usersService.getAllUsers();
-  }
-
   @Patch(':id')
-  updateUser(@Param('id') id: string, @Body() data: UpdateUserDto) {
+  updateUser(
+    @Param('id') id: string,
+    @Body() data: UpdateUserDto,
+    @GetUser('userId') userId: string,
+  ) {
+    if (id !== userId) {
+      throw new ForbiddenException('Access denied');
+    }
     return this.usersService.updateUser(id, data);
   }
 
   @Delete(':id')
-  deleteUser(@Param('id') id: string) {
+  deleteUser(@Param('id') id: string, @GetUser('userId') userId: string) {
+    if (id !== userId) {
+      throw new ForbiddenException('Access denied');
+    }
     return this.usersService.deleteUser(id);
   }
 }
